refactor(git): split ref resolution owner
